fix: connect linktree
