+++ src/pages/CreateLinktree.js
@@ -15,19 +15,8 @@ import {
   Textarea,
   Spacer,
   Spinner,
-  Card,
-  CardHeader,
-  CardBody,
-  CardFooter,
   Image,
   Stack,
-  Heading,
-  Divider,
-  ButtonGroup,
-  Checkbox,
-  Tooltip,
-  Select,
-  Center,
   Radio,
   RadioGroup,
 } from "@chakra-ui/react";
@@ -41,8 +30,7 @@ import "../css/ButtonAnimations.css";
 
 function makeStorageClient() {
   return new Web3Storage({
-    token:
-      "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJkaWQ6ZXRocjoweDhmOWMxOTNjODJlODMzMjVDMThkNWM4NzRCM2Q2NGM5ZjI5NDdEOUQiLCJpc3MiOiJ3ZWIzLXN0b3JhZ2UiLCJpYXQiOjE2ODM2NTY1NzExNjEsIm5hbWUiOiJLb2lpIn0.qZJmInvmwLCkq_7T3h2gfm4Hs84MNKEVooOuAFfbIXI",
+    token: process.env.REACT_APP_WEB3STORAGE_TOKEN,
   });
 }
